This python file is made by me for finding the alternative of the conventional IRAF method of doing the photometry….
Limitations:
this code can find the apparent magnitude of different filters but it can't calculate the true magnitude of the different filters. We conventionally use the extinction, coefficient method to determine the true magnitude uing the fitting method. but this can't be done using this python method. So, I am leaving this method mid way, but it is working fine till finding the apparent magnitude for each filters…